+++ Changelog.md
@@ -1,5 +1,8 @@
 ## Gazebo 7.0
 
+1. Use opaque pointers for the rendering/Heightmap class.
+    * [Pull request #](https://bitbucket.org/osrf/gazebo/pull-request/)
+
 1. Use opaque pointers for the Master class.
     * [Pull request #2036](https://bitbucket.org/osrf/gazebo/pull-request/2036)